samples/analysis_server/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pillow==11.3.0
numpy==2.3.2
ultralytics==8.3.173
python-multipart==0.0.20
websockets==15.0.1
